+++ src/TreeHouse/Queue/Amqp/Driver/Amqp/Publisher/AmqpMessagePublisher.php
@@ -2,11 +2,10 @@
 
 namespace TreeHouse\Queue\Amqp\Driver\Amqp\Publisher;
 
-use TreeHouse\Queue\Amqp\Driver\Amqp\AmqpFactory;
 use TreeHouse\Queue\Amqp\ExchangeInterface;
-use TreeHouse\Queue\Amqp\QueueInterface;
 use TreeHouse\Queue\Message\Composer\MessageComposerInterface;
 use TreeHouse\Queue\Message\Message;
+use TreeHouse\Queue\Message\MessageProperties;
 use TreeHouse\Queue\Message\Publisher\MessagePublisherInterface;
 
 class AmqpMessagePublisher implements MessagePublisherInterface
@@ -20,13 +19,6 @@ class AmqpMessagePublisher implements MessagePublisherInterface
      * @var MessageComposerInterface
      */
     protected $composer;
-
-    /**
-     * Automatically created queue for deferred messages.
-     *
-     * @var QueueInterface
-     */
-    protected $deferredQueue;
 
     /**
      * @param ExchangeInterface        $exchange
@@ -55,21 +47,13 @@ class AmqpMessagePublisher implements MessagePublisherInterface
     public function publish(Message $message, \DateTime $date = null, $flags = ExchangeInterface::NOPARAM)
     {
         if ($date instanceof \DateTime) {
-            $seconds = $date->getTimestamp() - time();
-
-            if ($seconds < 0) {
+            $delay = $date->getTimestamp() - time();
+            if ($delay < 0) {
                 throw new \OutOFBoundsException('You cannot publish a message in the past');
             }
 
-            if ($message->getRoutingKey()) {
-                // since we're using the routing key for the deferred queue, we cannot use it here
-                throw new \LogicException('Publishing delayed messages with a routing key is unsupported at the moment');
-            }
-
-            // publish with the routing key set to the deferred queue
-//            $flags = ExchangeInterface::NOPARAM;
-//            $message->setRoutingKey($this->getDeferredQueueName());
-//            $message->getProperties()->set('ttl', $seconds);
+            // set delay in milliseconds
+            $message->setHeader(MessageProperties::KEY_DELAY, $delay * 1000);
         }
 
         $body = $message->getBody();
@@ -77,39 +61,5 @@ class AmqpMessagePublisher implements MessagePublisherInterface
         $props = $message->getProperties()->toArray();
 
         return $this->exchange->publish($body, $route, $flags, $props);
-    }
-
-    /**
-     * @return \AmqpQueue
-     */
-    protected function getDeferredQueue()
-    {
-        if (null === $this->deferredQueue) {
-            $name = $this->getDeferredQueueName();
-
-            $factory = new AmqpFactory();
-            $this->deferredQueue = $factory->createQueue(
-                $this->exchange->getChannel(),
-                $name,
-                QueueInterface::DURABLE,
-                [
-                    'x-dead-letter-exchange' => $this->exchange->getName(),
-                    'x-dead-letter-routing-key' => '',
-                    'x-message-ttl' => 600, // 10 minutes by default
-                ]
-            );
-
-            $this->deferredQueue->bind($this->exchange->getName(), $name);
-        }
-
-        return $this->deferredQueue;
-    }
-
-    /**
-     * @return string
-     */
-    protected function getDeferredQueueName()
-    {
-        return sprintf('%s.deferred', $this->exchange->getName());
     }
 }