+++ src/TreeHouse/Queue/Amqp/Driver/Amqp/Queue.php
@@ -180,7 +180,13 @@ class Queue implements QueueInterface
      */
     public function consume(callable $callback, $flags = null, $consumerTag = null)
     {
-        $this->delegate->consume($callback, self::convertToDelegateFlags($flags), $consumerTag);
+        $wrapper = function (\AMQPEnvelope $envelope) use ($callback) {
+            $callback(new Envelope($envelope));
+
+            return false;
+        };
+
+        $this->delegate->consume($wrapper, self::convertToDelegateFlags($flags), $consumerTag);
     }
 
     /**