+++ tests/TreeHouse/Queue/Tests/Processor/Retry/RetryProcessorTest.php
@@ -2,7 +2,9 @@
 
 namespace TreeHouse\Queue\Tests\Processor\Retry;
 
-use TreeHouse\Queue\Message\Message;
+use Mockery as Mock;
+use Mockery\MockInterface;
+use TreeHouse\Queue\Amqp\EnvelopeInterface;
 use TreeHouse\Queue\Processor\ProcessorInterface;
 use TreeHouse\Queue\Processor\Retry\RetryProcessor;
 use TreeHouse\Queue\Processor\Retry\RetryStrategyInterface;
@@ -14,7 +16,7 @@ class RetryProcessorTest extends \PHPUnit_Framework_TestCase
      */
     public function it_can_be_constructed()
     {
-        $processor = new RetryProcessor($this->getProcessorMock(), $this->getStrategyMock());
+        $processor = new RetryProcessor($this->createProcessorMock(), $this->createStrategyMock());
 
         $this->assertInstanceOf(RetryProcessor::class, $processor);
     }
@@ -24,15 +26,12 @@ class RetryProcessorTest extends \PHPUnit_Framework_TestCase
      */
     public function it_can_get_and_set()
     {
-        $processor = new RetryProcessor($this->getProcessorMock(), $this->getStrategyMock());
-
-        $processor->setCooldownTime(1000);
-        $this->assertEquals(1000, $processor->getCooldownTime());
+        $processor = new RetryProcessor($this->createProcessorMock(), $this->createStrategyMock());
 
         $processor->setMaxAttempts(5);
         $this->assertEquals(5, $processor->getMaxAttempts());
 
-        $processor2 = new RetryProcessor($this->getProcessorMock(), $this->getStrategyMock());
+        $processor2 = new RetryProcessor($this->createProcessorMock(), $this->createStrategyMock());
         $processor->setProcessor($processor2);
         $this->assertSame($processor2, $processor->getProcessor());
     }
@@ -42,55 +41,38 @@ class RetryProcessorTest extends \PHPUnit_Framework_TestCase
      */
     public function it_can_process_a_message()
     {
-        $inner = $this->getProcessorMock();
-        $strategy = $this->getStrategyMock();
+        $inner = $this->createProcessorMock();
+        $strategy = $this->createStrategyMock();
 
-        /** @var RetryProcessor|\PHPUnit_Framework_MockObject_MockObject $processor */
-        $processor = $this
-            ->getMockBuilder(RetryProcessor::class)
-            ->setConstructorArgs([$inner, $strategy])
-            ->setMethods(['retryMessage'])
-            ->getMock()
+        $processor = new RetryProcessor($inner, $strategy);
+        $inner->shouldReceive('process')->once()->andReturn(true);
+        $strategy->shouldReceive('retry')->never();
+
+        $envelope = $this->createEnvelopeMock();
+        $processor->process($envelope);
+    }
+
+    /**
+     * @test
+     */
+    public function it_retries_when_processor_throws_exception()
+    {
+        $inner = $this->createProcessorMock();
+        $strategy = $this->createStrategyMock();
+
+        $exception = new \Exception();
+        $processor = new RetryProcessor($inner, $strategy);
+
+        $strategy
+            ->shouldReceive('retry')
+            ->once()
+            ->with(any(EnvelopeInterface::class), 2, $exception)
+            ->andReturn(true)
         ;
+        $inner->shouldReceive('process')->once()->andThrow($exception);
 
-        $processor->expects($this->never())->method('retryMessage');
-        $inner->expects($this->once())->method('process')->will($this->returnValue(true));
-
-        $processor->process(new Message('test'));
-    }
-
-    /**
-     * @test
-     */
-    public function it_cannot_process_a_message()
-    {
-        $inner = $this->getProcessorMock();
-        $strategy = $this->getStrategyMock();
-
-        $processor = new RetryProcessor($inner, $strategy);
-
-        $strategy->expects($this->once())->method('retry')->will($this->returnValue(false));
-        $inner->expects($this->once())->method('process')->will($this->returnValue(false));
-
-        $result = $processor->process(new Message('test'));
-
-        $this->assertFalse($result, 'The ->process() method should return the value from the strategy');
-    }
-
-    /**
-     * @test
-     */
-    public function it_cannot_process_with_exception()
-    {
-        $inner = $this->getProcessorMock();
-        $strategy = $this->getStrategyMock();
-
-        $processor = new RetryProcessor($inner, $strategy);
-
-        $strategy->expects($this->once())->method('retry')->will($this->returnValue(true));
-        $inner->expects($this->once())->method('process')->will($this->throwException(new \Exception()));
-
-        $result = $processor->process(new Message('test'));
+        $envelope = $this->createEnvelopeMock(1);
+        $result = $processor->process($envelope);
 
         $this->assertTrue($result, 'The ->process() method should return the value from the strategy');
     }
@@ -101,33 +83,49 @@ class RetryProcessorTest extends \PHPUnit_Framework_TestCase
      */
     public function it_cannot_exceed_max_retries()
     {
-        $inner = $this->getProcessorMock();
-        $strategy = $this->getStrategyMock();
+        $inner = $this->createProcessorMock();
+        $strategy = $this->createStrategyMock();
 
-        $inner->expects($this->any())->method('process')->will($this->returnValue(false));
+        $inner->shouldReceive('process')->once()->andThrow(new \Exception());
 
         // create message for second attempt
-        $message = new Message('test');
-        $message->getProperties()->set(RetryProcessor::PROPERTY_KEY, 2);
+        $envelope = $this->createEnvelopeMock(2);
 
         $processor = new RetryProcessor($inner, $strategy);
         $processor->setMaxAttempts(2);
-        $processor->process($message);
+        $processor->process($envelope);
     }
 
     /**
-     * @return \PHPUnit_Framework_MockObject_MockObject|ProcessorInterface
+     * @return MockInterface|ProcessorInterface
      */
-    protected function getProcessorMock()
+    private function createProcessorMock()
     {
-        return $this->getMockForAbstractClass(ProcessorInterface::class);
+        return Mock::mock(ProcessorInterface::class);
     }
 
     /**
-     * @return \PHPUnit_Framework_MockObject_MockObject|RetryStrategyInterface
+     * @return MockInterface|RetryStrategyInterface
      */
-    protected function getStrategyMock()
+    private function createStrategyMock()
     {
-        return $this->getMockForAbstractClass(RetryStrategyInterface::class);
+        return Mock::mock(RetryStrategyInterface::class);
+    }
+
+    /**
+     * @param int|bool $attempt
+     *
+     * @return MockInterface|EnvelopeInterface
+     */
+    private function createEnvelopeMock($attempt = false)
+    {
+        $mock = Mock::mock(EnvelopeInterface::class);
+        $mock
+            ->shouldReceive('getHeader')
+            ->with(RetryProcessor::PROPERTY_KEY)
+            ->andReturn($attempt)
+        ;
+
+        return $mock;
     }
 }